+++ src/components/content/display-date/display-date-preview.vue
@@ -1,0 +1,15 @@
+<template>
+	<preview-wrapper>
+		<template #title>
+			display-date
+		</template>
+
+		<preview-section>
+			<template #title>
+				Default date formatting
+			</template>
+
+			<display-date date="2025-03-29" />
+		</preview-section>
+	</preview-wrapper>
+</template>